requirements.txt
beautifulsoup4>=4.10.0
fastapi>=0.68.0
python-dotenv>=0.19.0
requests>=2.26.0
tweepy>=4.10.0
uvicorn>=0.15.0
openai>=1.82.0
